TOOL: renames internal folder from transformer to tree
